fixed overflow of long (moved to common)

+++ chemodurov.artem/S5/create-args-map.cpp
@@ -3,8 +3,8 @@
 chemodurov::args_dic_t chemodurov::createArgsMap()
 {
   args_dic_t map;
-  map.insert({"ascending", &Map< int, std::string >::traverse_lnr});
-  map.insert({"descending", &Map< int, std::string >::traverse_rnl});
-  map.insert({"breadth", &Map< int, std::string >::traverse_breadth});
+  map.insert({"ascending", &Map< long, std::string >::traverse_lnr});
+  map.insert({"descending", &Map< long, std::string >::traverse_rnl});
+  map.insert({"breadth", &Map< long, std::string >::traverse_breadth});
   return map;
 }
